farms

Module accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

Structs

AddRewards

Generated client accounts for AddRewards.

DepositToFarmVault

Generated client accounts for DepositToFarmVault.

HarvestReward

Generated client accounts for HarvestReward.

IdlMissingTypes

Generated client accounts for IdlMissingTypes.

InitializeFarm

Generated client accounts for InitializeFarm.

InitializeFarmDelegated

Generated client accounts for InitializeFarmDelegated.

InitializeGlobalConfig

Generated client accounts for InitializeGlobalConfig.

InitializeReward

Generated client accounts for InitializeReward.

InitializeUser

Generated client accounts for InitializeUser.

RefreshFarm

Generated client accounts for RefreshFarm.

RefreshUserState

Generated client accounts for RefreshUserState.

RewardUserOnce

Generated client accounts for RewardUserOnce.

SetStakeDelegated

Generated client accounts for SetStakeDelegated.

Stake

Generated client accounts for Stake.

TransferOwnership

Generated client accounts for TransferOwnership.

Unstake

Generated client accounts for Unstake.

UpdateFarmAdmin

Generated client accounts for UpdateFarmAdmin.

UpdateFarmConfig

Generated client accounts for UpdateFarmConfig.

UpdateGlobalConfig

Generated client accounts for UpdateGlobalConfig.

UpdateGlobalConfigAdmin

Generated client accounts for UpdateGlobalConfigAdmin.

WithdrawFromFarmVault

Generated client accounts for WithdrawFromFarmVault.

WithdrawReward

Generated client accounts for WithdrawReward.

WithdrawSlashedAmount

Generated client accounts for WithdrawSlashedAmount.

WithdrawTreasury

Generated client accounts for WithdrawTreasury.

WithdrawUnstakedDeposits

Generated client accounts for WithdrawUnstakedDeposits.