farms

Module cpi

Modules

accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is an AccountInfo. This is useful for CPI.

Structs

Return

Functions

add_rewards

deposit_to_farm_vault

harvest_reward

idl_missing_types

initialize_farm

initialize_farm_delegated

initialize_global_config

initialize_reward

initialize_user

refresh_farm

refresh_user_state

reward_user_once

set_stake_delegated

stake

transfer_ownership

unstake

update_farm_admin

update_farm_config

update_global_config

update_global_config_admin

withdraw_from_farm_vault

withdraw_reward

withdraw_slashed_amount

withdraw_treasury

withdraw_unstaked_deposits