farms

Module instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Structs

AddRewards

Instruction.

DepositToFarmVault

Instruction.

HarvestReward

Instruction.

IdlMissingTypes

Instruction.

InitializeFarm

Instruction.

InitializeFarmDelegated

Instruction.

InitializeGlobalConfig

Instruction.

InitializeReward

Instruction.

InitializeUser

Instruction.

RefreshFarm

Instruction.

RefreshUserState

Instruction.

RewardUserOnce

Instruction.

SetStakeDelegated

Instruction.

Stake

Instruction.

TransferOwnership

Instruction.

Unstake

Instruction.

UpdateFarmAdmin

Instruction.

UpdateFarmConfig

Instruction.

UpdateGlobalConfig

Instruction.

UpdateGlobalConfigAdmin

Instruction.

WithdrawFromFarmVault

Instruction.

WithdrawReward

Instruction.

WithdrawSlashedAmount

Instruction.

WithdrawTreasury

Instruction.

WithdrawUnstakedDeposits

Instruction.